Refactor TorrentTask into a viewmodel
