+++ Core/ControlMessages.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using bittorrent.Models;
+using Data=bittorrent.Models.Data;
 
 namespace bittorrent.Core;
